Monsters learn techniques on being created.
This commit also adds a number of techs for testing purposes. So far, the techniques appear to work fairly well. The system for monsters deciding to use techniques is incredibly barebones and needs to be fleshed out (it should NOT depend on MAXTECH) but overall it works, and makes the game way more dynamic.