Initial commit for new monster spell system.
Here we go. The big one. The initial idea for this is derived from FIQHack and the advice of FIQ on using bitfields to store monster spells/techs, but it will probably end up looking slightly different in execution. This commit merely ports over vanillas spells. In the future, we will be adding a boatload of unique monster spells across different monster classes.