New monster spellcasting system.
Although it was not originally intended to, this system functions very similarly to monster spellcasting in FIQHack. Monster spells are pulled from existing spellbooks, and players can learn monster spells as well. This will be expanded in the future to allow monsters to learn spells from spellbooks, and cast a wide variety of unique spells.